+++ packages/dataviews/src/components/dataviews/index.tsx
@@ -53,7 +53,6 @@ type DataViewsProps< Item > = {
 	: { getItemId: ( item: Item ) => string } );
 
 const defaultGetItemId = ( item: ItemWithId ) => item.id;
-const defaultGetItemLevel = ( item: any ) => item.level;
 const defaultIsItemClickable = () => true;
 const EMPTY_ARRAY: any[] = [];
 
@@ -66,7 +65,7 @@ export default function DataViews< Item >( {
 	actions = EMPTY_ARRAY,
 	data,
 	getItemId = defaultGetItemId,
-	getItemLevel = defaultGetItemLevel,
+	getItemLevel,
 	isLoading = false,
 	paginationInfo,
 	defaultLayouts,