+++ packages/dataviews/src/types.ts
@@ -485,7 +485,7 @@ export interface ViewBaseProps< Item > {
 	data: Item[];
 	fields: NormalizedField< Item >[];
 	getItemId: ( item: Item ) => string;
-	getItemLevel: ( item: Item ) => number;
+	getItemLevel?: ( item: Item ) => number;
 	isLoading?: boolean;
 	onChangeView: ( view: View ) => void;
 	onChangeSelection: SetSelection;